:white_check_mark: Updating tests based on comments
 #361
